Move traceChangedFiles to separate file
