+++ node-src/tasks/prepare/index.test.ts
@@ -1,31 +1,12 @@
-import { traceChangedFiles as traceChangedFilesDep } from '@cli/turbosnap';
-import { access } from 'fs';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import TestLogger from '../../lib/testLogger';
-import { calculateFileHashes, traceChangedFiles } from './index';
-
-vi.mock('fs');
-vi.mock('@cli/turbosnap');
-vi.mock('../readStatsFile', () => ({
-  readStatsFile: () =>
-    Promise.resolve({
-      modules: [
-        {
-          id: '../__mocks__/storybookBaseDir/test.ts',
-          name: '../__mocks__/storybookBaseDir/test.ts',
-        },
-      ],
-    }),
-}));
+import { calculateFileHashes } from './index';
 
 vi.mock('../../lib/getFileHashes', () => ({
   getFileHashes: (files: string[]) =>
     Promise.resolve(Object.fromEntries(files.map((f) => [f, 'hash']))),
 }));
-
-const traceChangedFilesTurbosnap = vi.mocked(traceChangedFilesDep);
-const accessMock = vi.mocked(access);
 
 const environment = { CHROMATIC_RETRIES: 2, CHROMATIC_OUTPUT_INTERVAL: 0 };
 const log = new TestLogger();
@@ -34,64 +15,6 @@ const http = { fetch: vi.fn() };
 afterEach(() => {
   vi.restoreAllMocks();
   vi.resetAllMocks();
-});
-
-describe('traceChangedFiles', () => {
-  beforeEach(() => {
-    accessMock.mockImplementation((_path, callback) => Promise.resolve(callback(null)));
-  });
-
-  it('sets onlyStoryFiles on context', async () => {
-    const deps = { 123: ['./example.stories.js'] };
-    traceChangedFilesTurbosnap.mockResolvedValue(deps);
-
-    const ctx = {
-      env: environment,
-      log,
-      http,
-      options: {},
-      sourceDir: '/static/',
-      fileInfo: { statsPath: '/static/preview-stats.json' },
-      git: { changedFiles: ['./example.js'] },
-      turboSnap: {},
-    } as any;
-    await traceChangedFiles(ctx, {} as any);
-
-    expect(ctx.onlyStoryFiles).toStrictEqual(Object.keys(deps));
-  });
-
-  it('escapes special characters on context', async () => {
-    const deps = {
-      './$example-new.stories.js': ['./$example-new.stories.js'],
-      './+example-new.stories.js': ['./+example-new.stories.js'],
-      './example-(new).stories.js': ['./example-(new).stories.js'],
-      './example[[lang=language]].stories.js': ['./example[[lang=language]].stories.js'],
-      '[./example/[account]/[id]/[unit]/language/example.stories.tsx]': [
-        '[./example/[account]/[id]/[unit]/language/example.stories.tsx]',
-      ],
-    };
-    traceChangedFilesTurbosnap.mockResolvedValue(deps);
-
-    const ctx = {
-      env: environment,
-      log,
-      http,
-      options: {},
-      sourceDir: '/static/',
-      fileInfo: { statsPath: '/static/preview-stats.json' },
-      git: { changedFiles: ['./example.js'] },
-      turboSnap: {},
-    } as any;
-    await traceChangedFiles(ctx, {} as any);
-
-    expect(ctx.onlyStoryFiles).toStrictEqual([
-      String.raw`./\$example-new.stories.js`,
-      String.raw`./\+example-new.stories.js`,
-      String.raw`./example-\(new\).stories.js`,
-      String.raw`./example\[\[lang=language\]\].stories.js`,
-      String.raw`\[./example/\[account\]/\[id\]/\[unit\]/language/example.stories.tsx\]`,
-    ]);
-  });
 });
 
 describe('calculateFileHashes', () => {
